+++ src/site/markdown/change-log.md
@@ -1,6 +1,7 @@
 1.2.1
 -------------
 * [Github issue #4](https://github.com/Axway/grapes-maven-plugin/issues/4) Manage paths in pom module fields
+* [Github issue #3](https://github.com/Axway/grapes-maven-plugin/issues/3) Manage project pom file correctly
 * [Github issue #1](https://github.com/Axway/grapes-maven-plugin/issues/1) Remove useless artifact resolution